Compresssion file moved to shared lib.
